// nautilus-ionic/src/app/api/model/requestCommandType.ts
/**
 * nautilus-central-service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type RequestCommandType = 'APPROVE' | 'REJECT';

export const RequestCommandType = {
    Approve: 'APPROVE' as RequestCommandType,
    Reject: 'REJECT' as RequestCommandType
};