// nautilus-ionic/src/app/api/model/requestStatus.ts
/**
 * nautilus-central-service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type RequestStatus = 'PENDING' | 'APPROVED' | 'REJECTED' | 'CANCELED';

export const RequestStatus = {
    Pending: 'PENDING' as RequestStatus,
    Approved: 'APPROVED' as RequestStatus,
    Rejected: 'REJECTED' as RequestStatus,
    Canceled: 'CANCELED' as RequestStatus
};